requirements.txt
# base version of ansible
# ansible~=2.10
ansible-core~=2.15

# https://github.com/readthedocs/readthedocs.org/issues/9038
Jinja2<3.1

# required for k8s ansible module
openshift>=0.13.1
# kubernetes >= 12.0.0

# required for vmware automation
pyvim==3.0.2
pyvmomi==7.0

# required for aws automation
boto==2.49.0
boto3~=1.26

# required for dns lookups
dnspython==1.16.0

# required for linting
yamllint==1.28.0
ansible-lint==6.8.7

# required to parse versions with rosa module
packaging==21.0

# semver processing
semver==2.13.0

# ip math
netaddr==0.8.0

pyopenssl==23.3.0

# ocm modules
git+https://github.com/rh-mobb/ocm-python.git@0.1.0
